Added code to support abandoned file processing
The Upload URL and Crop URL processing files are perfect for processing uploaded and cropped files as required.  However, there is no support for files which are abandoned by a user in either the uploaded or cropped stages.  These can of course be removed manually or by a CRON job...  but why not add functionality here to deal with them before they pile up and to eliminate the possibility of manual removal or a CRON job from interfering with a user.  

I added such functionality to the croppic.js and made a sample processing file.  I modeled the code on the existing code to facilitate additions if needed.